requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
aiohttp==3.11.11
aiosqlite==0.20.0
jinja2==3.1.5
python-dotenv==1.0.1
apscheduler==3.10.4
curl_cffi>=0.14.0
